MMSUP-2990: iterateChangesetsEvent.setMinVersion, align changesets query paramerters
