+++ xyz-hub-service/src/main/java/com/here/xyz/hub/rest/SpaceBasedApi.java
@@ -23,6 +23,10 @@ import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
 import static io.netty.handler.codec.http.HttpResponseStatus.CONFLICT;
 import static io.netty.handler.codec.http.HttpResponseStatus.NOT_FOUND;
 
+import static com.here.xyz.hub.rest.ApiParam.Query.END_VERSION;
+import static com.here.xyz.hub.rest.ApiParam.Query.START_VERSION;
+import static com.here.xyz.models.hub.Ref.HEAD;
+
 import com.here.xyz.events.ContextAwareEvent.SpaceContext;
 import com.here.xyz.hub.config.BranchConfigClient;
 import com.here.xyz.hub.config.TagConfigClient;
@@ -94,16 +98,55 @@ public abstract class SpaceBasedApi extends Api {
     return context.pathParam(Path.SPACE_ID);
   }
 
-  protected Ref getRef(RoutingContext context) throws HttpException {
+  protected Ref getRef(RoutingContext context, boolean changesetApi) throws HttpException {
     final String version = Query.getString(context, Query.VERSION, null);
     final String versionRef = Query.getString(context, Query.VERSION_REF, version);
 
     try {
-      return new Ref(versionRef);
+      if( !changesetApi )
+       return new Ref(versionRef);
+
+      // changesetApi
+      if( versionRef != null ) // versionRef or version  specified
+      {
+        Ref ref = new Ref(versionRef);
+        return ref.isRange() ? ref : new Ref( new Ref(0), ref); //TODO: minVersion hardcoded as 0 or align with space minversion?
+      }
+
+      // deprecated startVersion/endVersion
+      long startVersion = getLongQueryParam(context, START_VERSION, 0);
+      long endVersion = getLongQueryParam(context, END_VERSION, -1);
+
+      if (endVersion != -1 && startVersion > endVersion)
+       throw new IllegalArgumentException("The parameter \"" + START_VERSION + "\" needs to be smaller than or equal to \"" + END_VERSION + "\".");
+
+      Ref ref = new Ref( new Ref(Math.max(0, startVersion - 1)) , endVersion == -1 ? new Ref(HEAD) : new Ref(endVersion) ); //TODO: minVersion hardcoded as 0 or align with space minversion?
+
+      return ref;
+
     }
     catch (InvalidRef e) {
       Map<String, String> placeholders = Map.of("versionRef", versionRef, "cause", e.getMessage());
       throw new DetailedHttpException("E318404", placeholders, e);
+    }
+  }
+
+  protected Ref getRef(RoutingContext context) throws HttpException {
+    return getRef( context, false);
+  }
+
+  private long getLongQueryParam(RoutingContext context, String paramName, long defaultValue) {
+    try {
+      long paramValue = Query.getLong(context, paramName);
+      if (paramValue < 0)
+        throw new IllegalArgumentException("The parameter \"" + paramName + "\" must be >= 0.");
+      return paramValue;
+    }
+    catch (NullPointerException e) {
+      return defaultValue;
+    }
+    catch (NumberFormatException e) {
+      throw new IllegalArgumentException("The parameter \"" + paramName + "\" is not a number.", e);
     }
   }
